central text fix

+++ app/src/main/java/br/gohan/qualiar/ui/MeterState.kt
@@ -2,7 +2,7 @@ package br.gohan.qualiar.ui
 
 //states model 
 class MeterState(
-    val polutionText: Float = 0F,
+    val polutionText: Int = 0,
     val maxMeterValue: Float = 0F,
     val arcValue: Float = 0f,
     val inProgress: Boolean = false,
